Revamp code from util to service
